+++ src/main/java/biblioteca_de_jogos/Controladores/emprestimos/Emprestimos_emprestimosController.java
@@ -32,8 +32,8 @@ public class Emprestimos_emprestimosController {
     }
 
     public void Clicar_ok(ActionEvent event){
-        int IDusuario = Integer.parseInt(IDusuarioField.getText().trim());
-        int Idjogo = Integer.parseInt(IDjogoField.getText().trim());
+        Long IDusuario = Long.parseLong(IDusuarioField.getText().trim());
+        Long Idjogo = Long.parseLong(IDjogoField.getText().trim());
 
         Usuario usuario = UsuariosControl.getInstance().buscarUsuario(IDusuario);
         Jogo jogo = JogosControl.getInstance().buscarJogo(Idjogo);
