+++ src/main/java/biblioteca_de_jogos/control/JogosControl.java
@@ -8,12 +8,12 @@ import biblioteca_de_jogos.classes.Jogo;
 import biblioteca_de_jogos.model.Jogos;
 
 public class JogosControl {
-	private int id;
+	private Long id;
 	private Jogos jogos;
 	private static JogosControl self = null;
 	
 	private JogosControl() {
-		this.id = -1;
+		this.id = -1L;
 		this.jogos = Jogos.getInstance();
 		JogosControl.self = this;
 	}
@@ -26,7 +26,7 @@ public class JogosControl {
 		return JogosControl.self;
 	}
 
-	public Jogo buscarJogo(int id) {
+	public Jogo buscarJogo(Long id) {
 		Jogo achado = null;
 		for (Jogo jogo: this.jogos.getJogos()) {
 			if (jogo.getId() == id) {
@@ -36,7 +36,7 @@ public class JogosControl {
 		return achado;
 	}
 
-	public Jogo criarJogo(String nome, String editor, String descricao, int tempoPartida, int minJogadores, int maxJogadores, int copias, int idCategoria) {
+	public Jogo criarJogo(String nome, String editor, String descricao, int tempoPartida, int minJogadores, int maxJogadores, int copias, Long idCategoria) {
 		Jogo jogo = null;
 		if (nome != null && editor != null && tempoPartida > 0 && minJogadores > 0 && minJogadores < maxJogadores && copias > 0) {
 			jogo = new Jogo(this.id, nome, editor, descricao, tempoPartida, minJogadores, maxJogadores, copias, idCategoria);
@@ -110,7 +110,7 @@ public class JogosControl {
 		return this.jogos.getJogos();
 	}
 
-	public List<Jogo> buscarPorCategoria(int idCategoria) {
+	public List<Jogo> buscarPorCategoria(Long idCategoria) {
 		List<Jogo> achados = new ArrayList<Jogo>();
 		for (Jogo jogo: this.jogos.getJogos()) {
 			if (jogo.getIdCategoria() == idCategoria) {
