+++ src/main/java/biblioteca_de_jogos/control/UsuariosControl.java
@@ -6,12 +6,12 @@ import biblioteca_de_jogos.classes.Usuario;
 import biblioteca_de_jogos.model.Usuarios;
 
 public class UsuariosControl {
-	private int id;
+	private Long id;
 	private Usuarios usuarios;
 	private static UsuariosControl self = null;
 	
 	public UsuariosControl() {
-		this.id = -1;
+		this.id = -1L;
 		this.usuarios = Usuarios.getInstance();
 	}
 
@@ -22,7 +22,7 @@ public class UsuariosControl {
 		return UsuariosControl.self;
 	}
 
-	public Usuario buscarUsuario(int id) {
+	public Usuario buscarUsuario(Long id) {
 		Usuario achado = null;
 		for (Usuario usuario: this.usuarios.getUsuarios()) {
 			if (usuario.getId() == id) {
